requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
requests
groq
python-dotenv